feat: migrate asset owner settings to TimelineManager and update AlbumOptionsModal
